pyproject.toml
[project]
name = "umai-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "~3.10"
dependencies = [
    "openai (>=1.60.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "gradio (>=5.13.2,<6.0.0)",
    "swarm (>=0.0.2,<0.0.3)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
